Refactor the update process into smaller chunks
