+++ redbot/_update/__main__.py
@@ -1,403 +1,15 @@
 import asyncio
-import os
-import shutil
-import sys
-import tempfile
-from operator import itemgetter
 from pathlib import Path
-from typing import Any, Final, Iterable, List, Literal, NoReturn, Optional, Set, Tuple
+from typing import Final, Optional, Tuple
 
 import click
-from packaging.specifiers import SpecifierSet
-from packaging.version import Version
-from python_discovery import PythonInfo, get_interpreter
-from rich.panel import Panel
-from rich.prompt import Confirm, IntPrompt, Prompt
-from rich.text import Text
 
-from redbot.core import data_manager
 from redbot.core._cli import asyncio_run
-from redbot.core.utils._internal_utils import fetch_latest_red_version
 
-from . import changelog, cog_compatibility_checker, common, runner
-from .tui import ChangelogReaderApp, ChangelogReaderResult
+from . import cmd, common, updater
 
 
-instance_data = data_manager.load_existing_config()
-if instance_data is None:
-    instance_list = []
-else:
-    instance_list = list(instance_data.keys())
-
-_EXIT_INSTANCE_SITE_PREFIX_MISMATCH: Final = 4
-_CHECK_COG_COMPATIBILITY_CMD_NAME: Final = "check-cog-compatibility"
-_RED_VERSION_CMD_ARG_NAME: Final = "--red-version"
-_PYTHON_VERSION_CMD_ARG_NAME: Final = "--python-version"
 _CHECK_OTHER_PYTHON_INSTALLS_CMD_ARG_NAME: Final = "--check-other-python-installs"
-
-
-def _get_system_interpreters(
-    requires_python: SpecifierSet,
-) -> List[Tuple[str, Version, PythonInfo]]:
-    interpreters = {}
-
-    def _append_interpreter(info: PythonInfo) -> Literal[False]:
-        version = Version(info.version_str)
-        if version in requires_python:
-            # realpath call is needed because get_interpreter lists
-            # /usr/bin and /bin as separate even though they're the same path
-            interpreters[os.path.realpath(info.executable)] = (version, info)
-        return False
-
-    get_interpreter("cpython", predicate=_append_interpreter)
-
-    ret = [(key, *value) for key, value in interpreters.items()]
-    ret.sort(key=itemgetter(1), reverse=True)
-    return ret
-
-
-def _search_for_interpreters(
-    requires_python: SpecifierSet,
-) -> List[Tuple[str, Version, PythonInfo]]:
-    console = common.get_console()
-    with console.status("Searching for compatible Python interpreters on your system..."):
-        interpreters = _get_system_interpreters(requires_python)
-
-    if not interpreters:
-        url = "https://docs.discord.red/en/latest/install_guides/"
-        console.print(
-            f"{common.ICON_ERROR} Could not find a compatible Python interpreter!\n"
-            'Please follow the steps from the "Installing the pre-requirements" section'
-            " of the install guide for your system:"
-        )
-        console.print(Text(url, style=f"link {url}"))
-        console.print("Once you finish installing the pre-requirements, run this command again.")
-        raise SystemExit(1)
-
-    return interpreters
-
-
-def _ask_for_interpreter(
-    *, current_python_version: Version, requires_python: SpecifierSet
-) -> Tuple[str, Version, PythonInfo]:
-    interpreters = _search_for_interpreters(requires_python)
-    console = common.get_console()
-
-    def _render_interpreter(interpreter_exe: str, interpreter_version: Version) -> Text:
-        return Text.assemble(
-            "CPython ",
-            (str(interpreter_version), "repr.number"),
-            " (",
-            (interpreter_exe, "log.path"),
-            ")",
-        )
-
-    text = Text("Found the following compatible Python interpreters on your system:")
-    for idx, (interpreter_exe, interpreter_version, python_info) in enumerate(interpreters, 1):
-        text.append_text(Text(f"\n{idx}. ", style="markdown.item.number"))
-        text.append_text(_render_interpreter(interpreter_exe, interpreter_version))
-    console.print(Panel(text))
-
-    while True:
-        result = IntPrompt.ask(
-            "\nEnter the number of the Python interpreter above that you want to use"
-            " or type 0 to input the path to it yourself. Generally, you should choose"
-            " the interpreter with the latest version on the above list.\nEnter your selection",
-            default=1,
-        )
-        if result < 0 or result > len(interpreters):
-            console.print("[prompt.invalid] This is not a valid choice.")
-            continue
-
-        if result == 0:
-            response = Prompt.ask(
-                "Please input the path to the Python interpreter that you want to use"
-            )
-            if not response:
-                console.print("[prompt.invalid] No path was provided.")
-                continue
-            info = PythonInfo.from_exe(response, raise_on_error=True)
-            interpreter_version = Version(info.version_str)
-            if info.implementation != "CPython" or interpreter_version not in requires_python:
-                console.print(
-                    "[prompt.invalid] The provided path points to an incompatible Python"
-                    f" interpreter. Latest version requires CPython {requires_python} but"
-                    f" the provided interpreter is {info.implementation} {interpreter_version}."
-                )
-                continue
-            interpreter_exe = info.executable
-        else:
-            interpreter_exe, interpreter_version, info = interpreters[result - 1]
-
-        console.print(
-            "\n[b]You selected:[/]", _render_interpreter(interpreter_exe, interpreter_version)
-        )
-        if Confirm.ask("Do you want to continue with this choice?"):
-            console.print()
-            break
-
-    return interpreter_exe, interpreter_version, info
-
-
-async def main(
-    *,
-    instances: List[str],
-    excluded_instances: Set[str],
-    ignore_prefix: bool,
-    backup_dir: Optional[Path],
-    no_backup: bool,
-) -> None:
-    console = common.get_console()
-    current_version = common.get_current_red_version()
-    current_python_version = common.get_current_python_version()
-
-    with console.status("Checking latest version..."):
-        latest = await fetch_latest_red_version()
-
-    if current_version >= latest.version:
-        common.print_with_prefix_column(
-            common.ICON_SUCCESS,
-            "You are already running the latest available version of Red.",
-        )
-        return
-
-    common.print_with_prefix_column(
-        common.ICON_SUCCESS,
-        Text("New version available: ").append(str(latest.version), style="bold"),
-    )
-
-    breaking_update = current_version.release[:2] != latest.version.release[:2]
-    with console.status("Fetching changelogs..."):
-        changelogs = await changelog.fetch_changelogs()
-        changelogs = changelog.get_changelogs_newer_than(changelogs, current_version)
-    common.print_with_prefix_column(common.ICON_SUCCESS, "Changelogs fetched.")
-
-    first_changelog_version = min(changelogs)
-    last_changelog_version = max(changelogs)
-    parts = []
-    if first_changelog_version == last_changelog_version:
-        parts.append(
-            "You will now be presented with the changelog for"
-            f" [b]Red {first_changelog_version}[/]."
-        )
-    else:
-        parts.append(
-            "You will now be presented with the changelogs for"
-            f" [b]Red {first_changelog_version}[/]-[b]{last_changelog_version}[/]."
-        )
-    parts.append(
-        f"\n[bold]{common.ICON_WARN}"
-        '  Make sure to read through the [green]"Read before updating"[/] section'
-        f" before continuing. {common.ICON_WARN}[/bold]\n"
-    )
-    if breaking_update:
-        parts.append(
-            f"[bold]{common.ICON_WARN}"
-            "  Please note that this is a major release and it may have some changes that your bot"
-            " or its cogs are are affected by.[/bold]\n"
-        )
-    parts.append(
-        "After the changelog is open and you're ready to continue, hit the [b]Q[/] key"
-        " to close the changelog and continue the update process.\n\n"
-        "Hit the [b]Enter[/] key to show to the changelog."
-    )
-    console.input(Panel("".join(parts)), password=True)
-
-    viewer = ChangelogReaderApp(changelog.render_markdown(changelogs))
-    result = await viewer.run_async()
-    if result is None:
-        raise RuntimeError("Unexpected state")
-    if result is ChangelogReaderResult.QUIT:
-        raise click.Abort()
-
-    console.print("Changelog has been closed.\n")
-
-    interpreter_version = current_python_version
-    if current_python_version not in latest.requires_python:
-        common.print_with_prefix_column(
-            common.ICON_WARN,
-            "The latest version of Red requires a different version (",
-            Text(str(latest.requires_python), style="bold"),
-            ") from the one that you are currently using (",
-            Text(str(current_python_version), style="bold"),
-            ")\nredbot-update will have to recreate the virtual environment"
-            " with a compatible version of Python.",
-        )
-        interpreter_exe, interpreter_version, interpreter_info = _ask_for_interpreter(
-            current_python_version=current_python_version, requires_python=latest.requires_python
-        )
-    else:
-        interpreter_info = PythonInfo.current_system()
-        interpreter_exe = interpreter_info.system_executable
-
-    checked_instances = {}
-    failed_instances = []
-    for instance_name in instances:
-        if instance_name in excluded_instances:
-            continue
-        exit_code, stdout = await _call_check_cog_compatibility_cmd(
-            instance_name,
-            red_version=latest.version,
-            python_version=interpreter_version,
-            ignore_prefix=ignore_prefix,
-            internal=True,
-            stdout=asyncio.subprocess.PIPE,
-        )
-        if exit_code != _EXIT_INSTANCE_SITE_PREFIX_MISMATCH:
-            if exit_code:
-                failed_instances.append(instance_name)
-                print(stdout, end="")
-                Text.assemble(
-                    "\N{UPWARDS ARROW} " * 3, "Failure for ", (instance_name, "bold"), " instance"
-                )
-                console.rule(
-                    Text.assemble(
-                        "\N{UPWARDS ARROW} " * 3,
-                        "Failure for ",
-                        (instance_name, "bold"),
-                        " instance above",
-                        " \N{UPWARDS ARROW}" * 3,
-                    ),
-                    style="red",
-                )
-            else:
-                checked_instances[instance_name] = stdout
-        if stdout:
-            console.print()
-    console.print()
-
-    if checked_instances:
-        for instance_name, stdout in checked_instances.items():
-            console.rule(Text(instance_name, style="bold"))
-            print(stdout, end="")
-        console.rule()
-
-    common.print_with_prefix_column(
-        common.ICON_INFO,
-        "Finished checking cog compatibility.",
-        (
-            "\nThe results for each of the checked instances are shown above."
-            if checked_instances
-            else ""
-        ),
-    )
-    if failed_instances:
-        common.print_with_prefix_column(
-            common.ICON_ERROR,
-            "Failure occurred while trying to check compatibility for following instances: ",
-            Text(", ").join(
-                Text(instance_name, style="bold") for instance_name in failed_instances
-            ),
-            "\nScroll above to find the errors.",
-        )
-    if not checked_instances:
-        common.print_with_prefix_column(
-            common.ICON_INFO,
-            "There were no",
-            (" other" if failed_instances else ""),
-            " instances to check cog compatibility for.",
-        )
-    console.print()
-
-    to_backup = [] if no_backup else [*checked_instances, *failed_instances]
-    if no_backup:
-        common.print_with_prefix_column(
-            common.ICON_INFO, "Will not make backups as --no-backup option was passed."
-        )
-    else:
-        common.print_with_prefix_column(
-            common.ICON_INFO,
-            "The following instances will be backed up before performing the update: ",
-            Text(", ").join(Text(instance_name, style="bold") for instance_name in to_backup),
-        )
-    if breaking_update:
-        console.print(
-            "[b]Remember that this is a major release and it may have some breaking changes"
-            " that the bot or its cogs may be affected by.[/]"
-        )
-    if not Confirm.ask(f"Do you want to continue with the update to [b]Red {latest.version}[/]?"):
-        return
-    console.print()
-
-    if no_backup:
-        console.print("Will not make backups as --no-backup option was passed.")
-    else:
-        await _make_backups(backup_dir, to_backup)
-
-    _update_with_fresh_venv(interpreter_exe, interpreter_info, latest.version)
-
-
-def _update_with_fresh_venv(
-    interpreter_exe: str, interpreter_info: PythonInfo, latest_version: Version
-) -> NoReturn:
-    console = common.get_console()
-    venv_dir = Path(sys.prefix)
-    backup_dir = venv_dir / common.OLD_VENV_BACKUP_DIR_NAME
-    try:
-        backup_dir.mkdir()
-    except FileExistsError:
-        console.print(
-            "Found that a partial backup of a virtual environment from a past failed update exists"
-            " at",
-            Text(str(backup_dir), style="bold"),
-            "\nThe update will not proceed to avoid overriding it. If you are certain that"
-            " you don't need to restore anything from it, remove it and try updating again.",
-        )
-        raise SystemExit(1)
-
-    old_executable = Path(sys.executable)
-    rel_executable = old_executable.relative_to(venv_dir)
-    new_executable = backup_dir / rel_executable
-    wrapper_exe = runner.get_wrapper_executable()
-
-    for path in venv_dir.iterdir():
-        if path == backup_dir or path == wrapper_exe:
-            continue
-        path.rename(backup_dir / path.name)
-
-    console.print()
-    runner.make_exec_request(
-        str(new_executable),
-        "reinstall",
-        # base executable for venv creation
-        interpreter_exe,
-        # venv dir
-        str(venv_dir),
-        # scripts path
-        interpreter_info.sysconfig_path("scripts", {"base": str(venv_dir)}),
-        # Red dependency specifier
-        common.get_red_dependency_specifier(latest_version),
-    )
-
-
-async def _make_backups(backup_dir: Optional[Path], instances: Iterable[str]) -> Path:
-    backup_dir = backup_dir or Path(tempfile.mkdtemp(prefix="redbot-update-backup-"))
-    console = common.get_console()
-    console.print("Backups will be created at:", Text(str(backup_dir), style="bold"))
-    with console.status("Making a backup of the venv directory..."):
-        venv_dir = Path(sys.prefix)
-        shutil.copytree(venv_dir, backup_dir / "redenv", symlinks=True)
-
-    failed = []
-    for instance_name in instances:
-        console.print("Making a backup of the", Text(instance_name, style="bold"), "instance...")
-        proc = await asyncio.create_subprocess_exec(
-            sys.executable, "-m", "redbot.setup", "backup", instance_name, str(backup_dir)
-        )
-        if await proc.wait():
-            failed.append(instance_name)
-
-    if failed:
-        common.print_with_prefix_column(
-            common.ICON_ERROR,
-            "The following instances failed during backup: ",
-            Text(", ").join(Text(instance_name, style="bold") for instance_name in failed),
-            "\nScroll above to find the errors.",
-        )
-        if not Confirm.ask("Do you want to continue with the update regardless?"):
-            raise SystemExit(1)
-
-    return backup_dir
 
 
 @click.group(invoke_without_command=True)
@@ -406,7 +18,7 @@ async def _make_backups(backup_dir: Optional[Path], instances: Iterable[str]) ->
     "--include-instance",
     "included_instances",
     multiple=True,
-    type=click.Choice(instance_list),
+    type=click.Choice(common.INSTANCE_LIST),
     help="The list of instances to backup and check cog compatibility for. If not specified,"
     " all instances that use the current virtual environment will be backed up and checked.",
 )
@@ -414,7 +26,7 @@ async def _make_backups(backup_dir: Optional[Path], instances: Iterable[str]) ->
     "--exclude-instance",
     "excluded_instances",
     multiple=True,
-    type=click.Choice(instance_list),
+    type=click.Choice(common.INSTANCE_LIST),
     help="Exclude an instance from the list of instances to backup"
     " and check cog compatibility for.",
 )
@@ -472,16 +84,16 @@ def cli(
             # de-duplicate with order intact
             instances = list(dict.fromkeys(included_instances))
         else:
-            instances = instance_list
-        asyncio_run(
-            main(
-                instances=instances,
-                excluded_instances=set(excluded_instances),
-                ignore_prefix=ignore_prefix,
-                backup_dir=backup_dir,
-                no_backup=no_backup,
-            )
+            instances = list(common.INSTANCE_LIST)
+        options = updater.UpdaterOptions(
+            instances=instances,
+            excluded_instances=set(excluded_instances),
+            ignore_prefix=ignore_prefix,
+            backup_dir=backup_dir,
+            no_backup=no_backup,
         )
+        app = updater.Updater(options)
+        asyncio_run(app.run())
     # these should not be available to subcommands
     elif included_instances:
         raise click.NoSuchOption("--include-instance", ctx=ctx)
@@ -493,185 +105,7 @@ def cli(
         raise click.NoSuchOption("--no-backup", ctx=ctx)
 
 
-class VersionParamType(click.ParamType):
-    name = "version"
-
-    def convert(
-        self, value: Any, param: Optional[click.Parameter], ctx: Optional[click.Context]
-    ) -> Version:
-        if isinstance(value, Version):
-            if len(value.release) < 2:
-                self.fail(
-                    f"{value!r} needs to have at least 2 release components (major and minor).",
-                    param,
-                    ctx,
-                )
-            return value
-
-        try:
-            return self.convert(Version(value), param, ctx)
-        except ValueError:
-            self.fail(f"{value!r} is not a valid version number", param, ctx)
-
-
-@cli.command(_CHECK_COG_COMPATIBILITY_CMD_NAME)
-@click.argument(
-    "instances",
-    nargs=-1,
-    type=click.Choice(instance_list),
-    default=None,
-    metavar="[INSTANCE_NAME]",
-)
-@click.option(
-    _RED_VERSION_CMD_ARG_NAME,
-    type=VersionParamType(),
-    default=None,
-    help="The Red version to check cog compatibility for."
-    " If not provided, the information about latest available version will be fetched"
-    " and the command will check whether installed cogs support that version.\n"
-    "If this option is provided, --python-version also has to be provided.",
-)
-@click.option(
-    _PYTHON_VERSION_CMD_ARG_NAME,
-    type=VersionParamType(),
-    default=None,
-    help="The Python version to check cog compatibility for."
-    " If not provided, the command will either use the current interpreter's version or,"
-    " if that version is not compatible with the latest Red version, it will try to"
-    " find the latest available CPython interpreter on the system and will check whether"
-    " installed cogs support it.\n"
-    "If this option is provided, --red-version also has to be provided.",
-)
-@click.pass_context
-def check_cog_compatibility(
-    ctx: click.Context,
-    instances: Tuple[str, ...],
-    red_version: Optional[Version],
-    python_version: Optional[Version],
-) -> None:
-    """
-    Check if the installed cogs are compatible with the given version.
-    """
-    if (red_version, python_version).count(None) == 1:
-        raise click.BadParameter(
-            "Either both --red-version and --python-version options"
-            " have to be specified or neither.",
-            param_hint=[_RED_VERSION_CMD_ARG_NAME, _PYTHON_VERSION_CMD_ARG_NAME],
-        )
-
-    asyncio_run(
-        _check_cog_compatibility_command_impl(
-            red_version=red_version,
-            python_version=python_version,
-            instances=instances,
-            ignore_prefix=ctx.obj["IGNORE_PREFIX"],
-        )
-    )
-
-
-async def _check_cog_compatibility_command_impl(
-    *,
-    red_version: Optional[Version],
-    python_version: Optional[Version],
-    instances: Tuple[str, ...] = (),
-    ignore_prefix: bool = False,
-) -> None:
-    console = common.get_console()
-    if red_version is None or python_version is None:
-        with console.status("Checking latest version..."):
-            latest = await fetch_latest_red_version()
-            red_version = latest.version
-
-        python_version = Version(".".join(map(str, sys.version_info[:3])))
-        if python_version not in latest.requires_python:
-            interpreters = _search_for_interpreters(latest.requires_python)
-            _, python_version = interpreters[0]
-
-    if len(instances) == 1:
-        try:
-            await cog_compatibility_checker.check_instance(
-                instances[0],
-                latest_version=red_version,
-                interpreter_version=python_version,
-                ignore_prefix=ignore_prefix,
-            )
-        except cog_compatibility_checker.InstanceSitePrefixMismatchError as exc:
-            if not common.is_internal_cmd_call():
-                common.print_with_prefix_column(
-                    common.ICON_ERROR,
-                    Text(exc.instance_name, style="bold"),
-                    " instance could not be checked as it is a part of"
-                    " a different Python installation and/or virtual environment.",
-                )
-            raise SystemExit(_EXIT_INSTANCE_SITE_PREFIX_MISMATCH)
-        return
-
-    if not instances:
-        instances = tuple(instance_list)
-    checked_instances = []
-    for instance_name in instances:
-        exit_code, _ = await _call_check_cog_compatibility_cmd(
-            instance_name,
-            red_version=red_version,
-            python_version=python_version,
-            ignore_prefix=ignore_prefix,
-        )
-        if exit_code != _EXIT_INSTANCE_SITE_PREFIX_MISMATCH:
-            if exit_code:
-                raise SystemExit(exit_code)
-            checked_instances.append(instance_name)
-
-    if not checked_instances:
-        common.print_with_prefix_column(
-            common.ICON_ERROR, "There were no instances to check cog compatibility for."
-        )
-        raise SystemExit(1)
-
-
-async def _call_check_cog_compatibility_cmd(
-    instance_name: str,
-    *,
-    red_version: Version,
-    python_version: Version,
-    ignore_prefix: bool = False,
-    internal: bool = False,
-    stdout: Optional[int] = None,
-) -> Tuple[int, Optional[str]]:
-    args = [
-        "-m",
-        "redbot._update",
-        _CHECK_COG_COMPATIBILITY_CMD_NAME,
-        instance_name,
-        _RED_VERSION_CMD_ARG_NAME,
-        str(red_version),
-        _PYTHON_VERSION_CMD_ARG_NAME,
-        str(python_version),
-    ]
-    if ignore_prefix:
-        args.append(_CHECK_OTHER_PYTHON_INSTALLS_CMD_ARG_NAME)
-    env = os.environ.copy()
-    if internal:
-        env[common.INTERNAL_CMD_CALL_ENV_VAR] = "1"
-
-    # terminal woes
-    console = common.get_console()
-    if console.is_terminal:
-        env["TTY_COMPATIBLE"] = "1"
-        # Rich only checks stdout for Windows console features:
-        # https://github.com/Textualize/rich/blob/fc41075a3206d2a5fd846c6f41c4d2becab814fa/rich/_windows.py#L46
-        env[common.INTERNAL_LEGACY_WINDOWS_ENV_VAR] = "1" if console.legacy_windows else "0"
-    else:
-        # Rich does not set legacy_windows correctly when is_terminal is False
-        # https://github.com/Textualize/rich/issues/3647
-        env[common.INTERNAL_LEGACY_WINDOWS_ENV_VAR] = "0"
-    env["PYTHONIOENCODING"] = sys.stdout.encoding
-
-    proc = await asyncio.create_subprocess_exec(sys.executable, *args, env=env, stdout=stdout)
-    stdout_data, _ = await proc.communicate()
-    decoded_stdout = None
-    if stdout_data is not None:
-        decoded_stdout = stdout_data.decode()
-    return await proc.wait(), decoded_stdout
+cli.add_command(cmd.cog_compatibility.check_cog_compatibility)
 
 
 if __name__ == "__main__":
