+++ redbot/_update/cmd/__init__.py
@@ -1,0 +1,3 @@
+from . import cog_compatibility
+
+__all__ = ("cog_compatibility",)
