+++ redbot/_update/cmd/arg_names.py
@@ -1,0 +1,5 @@
+from typing import Final
+
+RED_VERSION: Final = "--red-version"
+PYTHON_VERSION: Final = "--python-version"
+CHECK_OTHER_PYTHON_INSTALLS: Final = "--check-other-python-installs"
